requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0

# Token counting
tiktoken>=0.5.0

# HTTP client
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0

# OpenAI (optional - for OpenAI models)
openai>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0